+++ src/common/SquareCollider.cpp
@@ -8,7 +8,7 @@
 
 #include "../headers/SquareCollider.h"
 
-SquareCollider::SquareCollider(double x, double y, double angle, double sideLength) {
+SquareCollider::SquareCollider(int x, int y, int angle, int sideLength) {
     this->_position = new Position(x, y, angle);
     this->_colliderType = ect_square_collider;
     this->_sideLength = sideLength;
@@ -18,10 +18,10 @@ bool SquareCollider::CheckCollision(ICollider * collider) {
     return false; // TODO
 }
 
-double SquareCollider::GetSideLength() {
+int SquareCollider::GetSideLength() {
     return this->_sideLength;
 }
 
-void SquareCollider::SetSideLength(double SquareCollider) {
+void SquareCollider::SetSideLength(int SquareCollider) {
     this->_sideLength = SquareCollider;
 }
